Fix for indentation/logic error in Ion.py
Logic/indentation error in the __init__ method of the class ion lead to the temperature being undefined (despite having been provided in the instantiation) when the ion was not in the masterlist. Also the message about not existing in the master list was printed when setup was False rather than when the ion was missing from the masterlist.